Changes the list paginator to consistently work off page numbers
This will allow us to work with just one number system instead of two.
There are some changes that need to be made in the view to help make
this easier. Make each page button be it's own object and we can then
access those through page numbers too. This will give our pager more
flexibility.

Amos and Kyle
[endlessm/eos-features-issues#707]