Don't generate multiple instances of the same .desktop file when filling desktop for QA
For now, we don't support multiple instances of the same .desktop file, so I had to update the QA script to avoid creating duplicates.  This is by design.  If, in the future, we want to allow multiple instances, we will need to revisit the logic for the shuffling animation, and can modify the QA script accordingly to intentionally create the duplicates.