Complete autotools gettext support
The framework was mostly in place but not quite complete,
and we were still using old scripts to generate the translation files.
The missing pieces have been added, and the old scripts removed.
Note that "make check" autogenerates the .pot file,
so the .pot file can also be removed from the repo.

[endlessm/eos-shell#532]