Avoid duplicates while filling desktop
The system does not support any duplicate items on the desktop
or in the desktop folders.

In the eos-fill-desktop script, we intended to not create copies
of any application or link that is already installed.
However, the existing logic did not prevent adding an item to the
desktop that exists in a folder.

[endlessm/eos-shell#806]