Allow selecting the app store page
We need to be able to select the page of the app store on start up,
using a command line switch; or remotely, using the DBus interface.

The three command line switches are:

  --apps, -a
        Selects the Applications page

  --folders, -f
        Selects the Folders page

  --web-links, -w
        Selects the Web page

The com.endlessm.AppStore.ShowPage() DBus method has also been added;
its argument is a string with the name of the page: 'apps', 'folders',
and 'web'.

[endlessm/eos-shell#925]