Add 17px spacing between rows in weblinks list
PListBox/GtkListBox does not support adding spaces between cells. Even trying
to add a margin does not work very well, as margin is added inside the row
itsef, and we do not have a way to specify in the CSS the first/last cell
should not have such margin.

So the selected approach is adding a invisible Frame as header of each row
(except for the first one), simulating thus a space.

A proper approach would be patching PListBox/GtkListBox to add support either
for spacing property, or fixing the use of margin property.

[endlessm/eos-shell#1055]