Handle correct states for 'Add' button in weblinks
'Add' button by default in disable state. When hovering over the row, button
becomes active. When hovering over the button, it becomes hovered, and when
clicked on it it becomes clicked. On being insensitive, the button becomes
insensitive.

Due some nasty bug due GTK+ 3.8 and PListBox (based on GtkListBox, from GTK+
3.10), the state of the row is propagated to the button itself; so if we click
on any place of the row, the button gets the 'pressed' state.

So we have used a workaround and create independent active/hover states for the
button, which are activated from the code.

[endlessm/eos-shell#1054]