Restore the content string extraction rules and files
The distcheck failure is a combo failure of both automake and intltool.
Sadly, those are the tools that we have.

We generate the json.h files that hold the translatable strings and put
them into the dist; this way intltool can generate the message catalogue
from $(srcdir), and automake does not break trying to generate files
inside $(srcdir) during distcheck.

It's a bit of autotools Dark Magic™, but I left a comment in the source
that ought to explain it.

[endlessm/eos-shell#1176]