Results of running unzip_content.py
This is run on the latest content in the CMS,
which adds the following apps:
- empathy
- frozen-bubble
- inkscape

It also updates several images that were previously
added by developers as placeholders, now using the
actual images provided by the designers.

Note that screenshots are not yet correct due to issues
with the CMS generated zip file.

[endlessm/eos-shell#1299]