Override preferred height request in TwoLinesLabel
We cannot use set_size_request() in the size_allocate() implementation:
we're effectively changing the size requisiont during the allocation
phase, which means that the label queues a relayout in the middle of a
relayout.

Instead, we should change the way we request the height for the given
width, and apply the same change during allocation.

[endlessm/eos-shell#1881]