Use AppManager.GetUserCapabilities()
If the user cannot install, update, or remove applications through the
application manager service, we still want them to allow adding and
removing shortcuts from the launchers grid.

In order to do that, we ask the AppManager for the capabilities of the
user, and only call the AppManager methods if the user satisfies the
minimum capabilities required.

[endlessm/eos-shell#1996]