Ensure the consistency of desktop id sets
When installing and uninstalling applications we need to ensure that the
sets of desktop ids held by the EosAppListModel are updated as well.

The sets will be blown away once the DBus API we call returns, but since
it's an asynchronous task, we need to keep the sets consistent in
between our methods and the remote ones.

[endlessm/eos-shell#1996]