app-list-model: fix logic for determining when an app is installed
An app is installed if GIO knows about it, not if its ID is in the list
reported by the shell. That is because we typically will have
applications that are not actually installed in there; also all the
other logic in the app store assumes that an installed app might or
might not have a launcher.

[endlessm/eos-shell#1996]