app-list-model: check for unprefixed launcher desktop IDs too
The shell icon grid always reasons in terms of unprefixed (i.e. without
leading 'eos-app-' desktop IDs), and so does the app store.
In order to avoid false negatives in the launcher checks, just look for
the unprefixed desktop ID first when searching for a launcher, falling
back to the prefixed/localized version only when that is not found.

This fixes double launchers being added for core applications installed
and on the desktop by default, such as the browser and the file manager.

[endlessm/eos-shell#2963]