app-list-model: change the way we load the model
Ensure the model is always loaded; this means that we trigger a thread
when the object is first constructed, and keep track of the changes as
they come. There's now no need to load the model explicitly anymore, nor
the risk to call into an unloaded model.

While at it, also add some better error checking during the load step.

[endlessm/eos-shell#3101]