app-list-model: remove locking
Take a step back on the mutex locking we do here; we hit a code path
where we block the UI waiting for a DBus method to return, which can
potentially be a long time since the app manager won't return from e.g.
ListAvailable() when there's another transaction in flight.

We can do better though, and make sure we only ever access the hash
tables from the main thread to avoid any issues.

[endlessm/eos-shell#3066]