Resize screenshots for narrow windows
If the window width is less than 1024, resize the screenshots
to leave the same width available for descriptions as for 1024.
It is especially important that the actual window width not
exceed the available screen area, as that pushes the close button
off-screen.

Since we don't really support window widths < 640,
let's use this as a threshold to ensure that we don't
shrink the screenshots too small.

[endlessm/eos-shell#3106]