Change app manager list applications from async to sync
This works around what appears to be a bug in the app manager,
where a sync call to ListApplications fails if there is an
outstanding async call to ListInstalled in progress.

If ListApplications really is desired to be async,
this commit can be reverted once the underyling issue
in the app manager is resolved.

[endlessm/eos-shell#3191]