appStoreWindow: Ellipsize and align subtitles if longer than two lines
Descriptions were taking too much space in some cases, instead of just
using two lines as maximum and ellipsizing the remaining text if needed.
Additionally, the title for each app was presented at different height
for cases with one-line and two-line descriptions, which is wrong too.

This patch forces the app cell to ellipsize the description if it's
longer than two lines and also uses a GtkStack with a dummy (invisible)
label to make sure that the visible description is always rendered as if
it took too lines, regardless of whether its that long or not.

[endlessm/eos-shell#3231]