weblinkListBoxRow: Don't change the style when moving the mouse over the row
The previous implementation higlighted the row by changing its style
and replacing the thumbnail with a desktop icon, restoring everything
when the row was no longer selected (mouse over it).

This patch removes all the code involved in performing those changes,
and comments out (temporarily) the call to the alternative "draw"
callback, which was responsible of rendering the desktop icon.

[endlessm/eos-shell#3233]