Reduce the left/right 'padding' inside install/remove buttons by 8px
It is not possible to reduce those 8px by just reducing the padding down
to 0px on each side because the current value for padding is just 4px.
Unfortunately, we can't just reduce the border width because we need to
consider the source image those borders are being sliced from, which
would need to be slightly stretched too, causing some distortion.

This patch implements a combination of reducing the padding down to 0px,
cropping the source images from 38x38 down to 36x38 (removing two fully
transparent 1px columns on each side), to reduce distortion, and then
stretching the images for the left and right borders, from 17px to 14px.

[endlessm/eos-shell#3494]