Replace _updateUrlFrameForState() with a new _setSiteUrlFrameChild() function
This change makes the patch smaller and easier to read. Additionally,
renamed _buildSearchBoxForNewSites into _buildWidgetsForSiteUrlFrame()
since now that helper function won't add any widget to the frame, but
just will build the GtkEntry and the GtkLabel widgets that will be shown
in that part of the UI, depending on the state.

[endlessm/eos-shell#3242]