EosLinkRow: Use a background image behind the icon for the installed link
Just using background-color for the GtkImage with the icon here is not
enough since the whole row is rendered in a special way, not just using
the UI file and CSS. If we do just CSS here, the dark background for the
icon will overflow the link's row, causing a weird effect in the borders.

So, we need to use the same technique to render the dark background
that should be shown under an installed icon, so the background looks
similar to the thumbnails we can see for every link that has not been
installed just now.

Note: I added a new asset (90x90 square filled with alpha(black, 0.20))

[endlessm/eos-shell#3233]