Fix icons on install button
Previously, there was an incorrect attempt to set the "install" class,
when install was the default.  Instead, we need a "launch" class
for the "Open app" button.  Also, the logic needs to take into account
whether or not the app is on the desktop.

For now, use placeholder assets for the "Open app" button icon
until assets are provided by the design team.
(In any case, the placeholder is more appropriate than the "+" icon.)

[endlessm/eos-shell#4010]