Change update option behavior
For the case where an application is installed on the device
but has an available update, change the option from "Install"
to "Add to desktop".  The justification is that users do not
necessarily want to use up bandwidth to update an app that
they already have available on the system.  Once the app has
been added to the desktop, the "Update" option is available.

[endlessm/eos-shell#4010]