appStoreWindow: flip window gravity on RTL
The app store window is special in that it extends out from a screen
edge. When the content in the window is too large and the window needs
to be resized, the window gravity determines the direciton the new size
will be extended towards.
In case we're using an RTL layout, use an eastern gravity instead of the
western default, so that the window will extend in the right direction
whenever the content is larger than the current allocation.

[endlessm/eos-shell#4051]