Avoid sliding in a half-drawn frame
Since the app store has a transparency, when it is hidden
the frame buffer retains the clone of the background on top
of which the semi-transparent app store content is drawn.
Since the processing of the content takes a while,
the app store slides in before the redraw completes,
and it appears that a copy of the desktop slides in ahead
of the app store.  Newer versions of X have a feature that
allows retaining the buffer contents of the hidden window,
but we do not currently support this version of X.

For now, we temporarily remove the contents from the window
upon showing so that it can be quickly drawn with just a
gray panel before the slide in animation.  Then, once
this gray panel is drawn, we can add the contents back
to the window and allow it to be redrawn.  The net effect
is that the window slides in with a semi-transparent plain gray
panel and then redraws once the app tile content is ready.

[endlessm/eos-shell#1895]