app-list-model: don't check for launcher presence before adding to shell
The shell is the ultimate arbiter on where a launcher goes in case it's
already specified in the layout. In fact, it already handles that case
just fine, appending it to the desktop layout as desired.
Just remove this check, since we don't want an icon that was previously
inside a folder, or is in a folder in the default layout, to end up in
that folder again when it's later added from the store.

[endlessm/eos-shell#3627]