Added ability to retry downloading of files for a short time
This change adds retry functionality to downloads so that if there's a
download failure, the logic will retry every X seconds for a maximum of
Y seconds. If any data is able to be retrieved, the retry counters are
reset so that we don't consider most temp interruptions in a global
context.

For now, X = 20 and Y = 4