info: Update existing AppInfo structure
We want to use the data from the installed bundles and the server to
update the EosAppInfo structures generated by the content data we load.

We already create the content of the app store using the curated data
coming from the eos-shell-content repository; there's really no point in
fetching information for applications that are not going to be shown, so
we simply use the EosAppInfo data structure as a base, and we update the
ancillary fields when needed.

[endlessm/eos-shell#4465]