Check for localized apps in any language
If not found in the user's language, check all supported languages
(starting with English).  For instance, if a user changes his/her
language, he/she should still be able to access the apps installed
on the system via the app store.

Note that we still do not check all languages when looking for apps
to download from the server.

[endlessm/eos-shell#3155]

Conflicts:
	EosAppStore/lib/eos-app-list-model.c