appFrame: fix page switch when updating or removing an app
We used to switch back to the grid in that case, as a consequence of a
too greedy invalidation. Now we only really should show the spinner
while asynchronously refreshing the model - everything else should
always be up to date in the model and follow the "changed" signal.

So we make spinning an explicit property of the page, and set it while
we refresh.

[endlessm/eos-shell#4854]