model: Correctly propagate errors when reloading
When reloading the model we check files locally and download them from
remote servers; both operations may fail for internal issues, and for
external factors. While we want the former to just end up in the Journal
for debugging purposes, we want to notify the user of the latter class
of errors.

We have all the infrastructure in place for doing so during operations
like installation and removal, we just need to do this for reloading
as well.

[endlessm/eos-shell#4537]