Shorter French translation of "Resources"
The original translation ("Documentations") caused tile layout
problems at 800x600.  The new translation ("Ressources") is based
on Google translate, and appears to be reasonable based on a
reverse translation.  In the future when we have full French
language support, we will want to review with a native speaker
who has the full context of the word's use within the app store.

[endlessm/eos-shell#4958]