Added partial handling code for parsing of updates meta record
We now invoke our skeleton json parsing function but since it's just a
skeleton, we still need to flesh it out. We also now have the monotonic
id stored in the list model so that we can reference the value as needed
though we may need to start it at -1 since the initial server value is
0 to ensure we get a clena update the first time (no locally downloaded
updates_meta.json).