Added retrieval of isDiff field from json record
We need this to decide how to process the update if it's newer than what
we have. There is a chance that we will need to move this to app info
module at some point since it's not quite a 'util' job to know the
details of th json record but we will explore that later.