Split out the "version" field into available and installed versions
Since we need to know if the delta can be installed based on the current
version we have, we can't keep overwriting this field hoping for the
best. This change should allow checking the fromVersion included in the
server JSON to figure out if the delta can be used or not.