Fixed issue where apps with newer deltas available couldn't be installed
Our logic was preventing newer deltas encountered from servers to be
overriden by full updates due to their lower version so it effectively
removed those apps from availabilty. We now don't create initial
app_info records with deltas for that reason but we now are unable to
upgrade to versions that have no matching updates but reworking that
would require keeping multiple "available_versions" which we don't want
to worry about right now. In the process, we also needed to ensure that
the first full update runs through the main logic to assign delta fields
(if needed) so the stored_version calculation was also moved up.