Fixed emission of last download progress event to the model
Old code would send the last download progress event to main context and
then free the struct from the model and by the time that the main
context would need the struct data, it would already be freed. Now we
pass the destroy notify function to the downloader so that it can free
the data also in the main context - it makes the net utils a bit more
complex but also more sturdy.