Unify order of parameters for AppStore's show() and doShow() functions
We were passing the 'reset' and 'timestamp' parameters swapped in order
in AppStoreDBusService, which was messing up with the value passed to
Gtk.Window.present_with_time later on in AppStoreWindow.showPage(),
which was interfering with the mouse events that were supposed to make
the Gtk.Entry widgets grab the focus.

This happened probably because the order of those parameters in expected
by AppStoreWindow.doShow() and the ones used from AppStore.show() were
swapped already, leading to some confusion. This patch fixes that, while
unifying the expected order of those parameters in these three functions.

[endlessm/eos-shell#5266]