Don't assume we always have all the deltas
Currently the app store code assumes that when deltas are enabled, we
have every single delta available on the server to update from arbitrary
bundle versions.
This assumption does not hold true; when we don't have a valid delta for
the version we have currently installed on the system, just fall back to
the full update.