folderFrame: Implement right behaviour when clicking on a folder icon
At the moment, a new bubble opens each time the user clicks on a folder
icon (potentially showing many of them at the same time) and the focus
is not automatically switched to the bubble's text entry when the user
starts typing, which is confusing.

This patch fixes does two issues at once, although it has to implement a
workaround for the automatic focus part due to a limitation in GTK+,
which does not allow moving the focus to a GtkEntry without removing the
placeholder text, when the user hasn't started typing yet.

[endlessm/eos-shell#5479]