appFrame: Always make sure to update the state after uninstalling
Normally, the state will be signalled from the app list model, which is
updated from the Shell and Gio, but in certain circumstances, it's
possible to "beat the race" and have Shell and Gio both poke our
callbacks asynchronously before the transaction is complete. Always
update at the end of a transaction to guarantee that our state will be
consistent.