Fix problems with server endpoint swappings
Previously we were setting the minimum versions that we cared about at
the currenly installed version which means that if the installed version
is above what the server can provide, uninstalling said app will cause
it to not be "available" and thus have no installation button visible.

New behavior does not filter out any records due to installed version so
we properly process server records that have versions below the ones on
the client. This increases the processing time a bit but it's important
enough that we are willing to do that tradeoff.