appFrame: fix a problem with scrolled windows
Currently the code always recreates a scrolled window when it populates
the page, but then only ever destroys its child. This makes it so that
every time the category is loaded, the content will be pushed down and a
new child will be added to the conten box.

Instead, always re-use the same scrolled window. At the same time move
it to the base class, as both subclasses need it.

[endlessm/eos-shell#5016]