Change how the app store loads and refreshes pages
Currently, we refresh the app store pages very often - basically every
time anything changes or we click on a sidebar item the whole state is
discarded.
Some of our pages are static (e.g. folder page), and others only should
be refreshed (or even loaded in the first place) when they're the
currently active page.
We now introduce a thin page manager interface, and only load the pages
we need on demand. We also take care of not invalidating more state than
we should when anything changes.

[endlessm/eos-shell#5724]