Change progress data ownership while downloading
Currently we pass around a destroy notify function to the methods in
eos-net-utils in order to get the progress data freed when the download
operation finishes.
This is not typically needed, as all the functions in there are
synchronous. EosAppListModel will call them from a separate thread, so
it is most logically correct to have it also bounce back to the main
context when needed.
But more importantly, EosAppListModel knows the content of the payload
that goes into the signal, so it can e.g. call g_object_ref() at the
appropriate moment.

This also fixes a crash noticeable on EC-100 while cancelling network
operations.

[endlessm/eos-shell#5792]