net-utils: do not send invalid resume ranges to the server
When we try to resume a file that we already have fully downloaded, the
server will respond back with an "invalid range" error.
Prevent this by avoiding download altogether when the file is already
the same length of the one on the server.
Also, do not try to resume files that are larger than the original.

[endlessm/eos-shell#5793]