appFrame: revalidate on _showView(), not reset()
In order to keep the sort stable after an operation was completed, we
were overriding the reset() method to invalidate the view later.
But that does not take into account the case when the view is returned
to from the details page.
Moving the same override in _showView() will take care of that.

[endlessm/eos-shell#5912]