app-list-model: don't call CancelTransaction() unless it was cancelled
Currently the app store calls CancelTransaction() whenever any error
occurs during the call to CompleteTransaction().

As the app manager already cleans up transactions that fail for internal
errors, the only case when it makes sense to call CancelTransaction() is
actually when the app store initiated the cancellation, i.e. if the
error matches G_IO_ERROR_CANCELLED.

This also fixes a potential crash that might occur when an internal error
happened in the app manager, causing it to free the transaction, and we
call back into CancelTransaction() causing the transaction to be freed
again.

[endlessm/eos-shell#5923]