app-list-model: do not assume info != NULL in _finish() methods
EosAppListModel still uses desktop IDs for install/uninstall/update
functions. This means that the desktop ID we are given could be
non-existant, in which case we will complete the task with an error.

When the caller calls into the _finish() function though, code assumes
that the task data was set to an EosAppInfo, which will not be the case
in the situation described above.

This commit adds NULL checks before calling into EosAppInfo's API.

[endlessm/eos-shell#5904]