Move critical error logging outside reusable function
When testing language names, it is expected that some will not match,
so we don't want to log all such errors as critical.  Instead, log as
debug inside the reusable function (where we have the file name),
and leave it to the caller to log the errors as critical if applicable.

[endlessm/eos-shell#5922]