style: adjust minimum scrollbar height for small screens
The current minimum scrollbar length is 115px, which is way
too big for composite mode. This is clearly visible in the
websites page, where the scrollbar covers the whole area,
making it impossible to scroll without a scrollwheel.

To fix that, simply set the minimum scrollbar height to 30px.
Values lower than that would probably break theming, and 30px
works well.

[endlessm/eos-shell#5925]