appInstalledBox: make stack non-hhomogeneous
When the GtkStack is hhomogeneous, it allocates
the same width for all it's children. In the specific
case of appInstalledBox, it was allocating too much
space for the controls box. In fact, for 720x480
resolution, it was making the toplevel window to
reallocate in order to make the stack's row fit.

To fix that, simply set GtkStack::hhomogeneous to False
and let each stack child have it's necessary space.

[endlessm/eos-shell#5925]