appInstalledBox: don't set width-chars
By setting width-chars, we're setting minimum label
sizes. In practice, that means that the app store won't
work when in composite mode, which requires a more
flexible layout.

Fix that by removing the width-chars and setting the
ellipsize property of the labels. Removing the width
chars won't affect anything else since the labels are
exanding.

[endlessm/eos-shell#5925]