mainWindow: make header title label ellipsizable
When dealing with small screens, the title label can't
be wraped since it'd eat up all available space. Also,
it can't resize the window, since the screen might be
too narrow for that.

In order to fix that, make it ellipsizable, so it shrinks
enought to fit the available window's width.

[endlessm/eos-shell#5925]