appInfoBox: fix logic for remove button visibility
Since the introduction of the installed apps page, we have changed
the logic for the remove button visibility not to take into account
whether a desktop launcher exists.
Fix a location where the logic wasn't updated, causing the remove
button not to be visible when an app had an update available.

[endlessm/eos-shell#6143]