weblinkFrame: do not wait forever if the URL does not provide a title
We were considering a website loaded as soon as the title was ready, so
that we would not have to wait until the whole document finishes loading
just to show to the user that a website has been found. The problem with
this approach is that we will keep waiting forever if the URL does not
provide a valid title, even if the whole document finishes loading.

To fix it, we better connect to the load-changed event from the webview
to make sure we always stop when the load is finished, regardless of the
title being available or not. If it's available, we will still mark the
URL as 'found' without waiting for the load to finish (to look more
responsive), otherwise we craft one ourselves based on the 'host' and
'path' parts of the URL, to show something that is not just the raw URL.

[endlessm/eos-shell#6178]