model: Check available space on update
When updating we want to keep the application on the same storage type,
unless there's not enough space to hold both the app and the backup we
use for rolling back in case of error.

We do a check on the app store side to avoid going through with the
update and then having to redo the whole update logic in case of error;
this is also needed because the app manager does not return a specific
"out of disk space" error (yet).

[endlessm/eos-shell#6136]