model: Cope with updates requiring more disk space
When performing updates, the app manager does a full back up of the
existing application; this may end up exhausting the available space on
the target prefix, which causes the update itself to fail.

The app store already does a check on the available disk space and
changes the target prefix when installing a new app; we need it also
when doing an update.

This requires an API change in com.endlessm.AppManager.

[endlessm/eos-shell#6136]