lib: Use the right path to check for free space
The eos_check_available_space() function takes a path to a file and then
extracts the parent component; if we pass the path of the storage prefix
then it's going to use the wrong file system when checking the available
free space.

There's also the matter of it hardcoding the minimum size to twice the
actual requested size, which is another fairly big gotcha.

Let's make the function a bit less clever, and just check the available
space at the given GFile, and using the requested minimum size; we
should move all the set up to the two call sites.

[endlessm/eos-shell#6136]