info-box: Don't hide store directly after launching app
When launching an app, if the store is hidden before the app's window
is visible (most cases), the app is shown as minimized because of what
seems to be a race between showing the app and showing the desktop.

These changes stop hiding the app store's window directly because it
will be hidden anyway when the app's window is shown.

Keeping the store visible until the app is shown will also allow to give
feedback to the user about what's being done (opening the app), which is
good for apps that take some time to be shown. This will however be
implemented on a different commit.

[endlessm/eos-shell#4190]