info-box: Inform the user when an app is being opened
When opening an app, nothing is shown after the "open" button is
clicked. This can lead to users clicking the button again if an app
is taking a little while to be opened.
These changes add information about the mentioned state, indicating
that the user should wait for the application to be launched.

[endlessm/eos-shell#4190]