appFrame: refactor search in a separate frame widget
Previosuly, search was wired up to be a feature of AppFrame. However,
AppFrame is a base class, and at any given time there are a handful of
AppFrame widgets (one for each category).
The result was that search was actually being computed many times in a
row in the model.

It's much simpler to move the search code to the window, and split the
UI part in its own category page. We take care of setting a NULL name
for it, so that it does not appear in the sidebar.

[endlessm/eos-shell#5019]