Remove all the app pages
As part of our transition from eos-app-store to GNOME Software, we no
longer allow installing apps from the app store. It will be used only to
handle weblinks and desktop folders, and only as a temporary transition
measure.

Accordingly, we need to modify the WeblinkFrame to wait for the
appListModel to finish loading before we use it, since we can't rely on
the AppPageProvider ensuring this anymore.

We also need to modify the default page.