weblinkFrame: Properly handle failed page loads
When a page load fails, WebKit emits load-failed and then transitions
to the LOAD_FINISHED state. Here we are assuming that LOAD_FINISHED
indicates a successful load, which is not correct, and causes us to
inappropriately transition away from the error state immediately after
every error. It's also causing a JS error.

https://phabricator.endlessm.com/T11604