Fix how OS details are checked in updates_meta.js
When the updates_meta.js does not exist or does not have the OS details
(version and personality), it is mistakenly not considering it as
invalid. This makes the logic that considers the apps list outdated to
work only when the updates_meta.js file is cached.

These changes check the OS details before checking is the file is cached
or downloading it, so it will always consider the file invalid unless
all its info is up to date.

https://phabricator.endlessm.com/T11597