weblinkFrame: Remove unnecessary D-Bus calls
The previous commit was the simplest way to remove WeblinkFrame's use of
AppListModel, but it introduced several unnecessary sync D-Bus calls to
determine if a weblink is on the desktop, one for each
WeblinkListBoxRow. Instead, have WeblinkFrame do this work once for all
rows, and pass the result on to the rows.

https://phabricator.endlessm.com/T12532