weblinkFrame: avoid flashing "added successfully" text
I introduced a minor visual glitch in T12532. After adding a weblink,
the description for that weblink is supposed to change to the text
"Added successfully" for the second or two before the app store hides
itself to show the desktop. But as part of T12532, I snuck in a change
to repopulate the app store when the desktop icons change, in case an
application is uninstalled; this causes the tooltip to immediately
switch away from "Added successfully" back to the original description
text.

Fix this by ignoring desktop changes during the 500ms period between
when a weblink is installed and when the app store hides itself.

https://phabricator.endlessm.com/T12794